feat: 기존 medicine.js 로직 변경 없이 책임 나누기
